requirements.txt
mediapipe
cvzone
pillow
tensorflow
scikit-learn
django>=4.0.2,<4.1.0
django-import-export
sqlalchemy
virtualenv
python-dotenv==0.21.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
gunicorn==20.1.0
whitenoise==6.2.0
Brotli==1.0.9